honest types
Was getting nervous about the idea of a val accidentally stumbling
upon the not_yet_full sentinel through some sort of situation using
the val pointer as a punned integer... but that's the sort of trick
we expect to play only with _keys_.

Switching key over to intptr_t (any value up to a pointer's size) and
leaving val as void* (this should be a real valid pointer) signals the
limits of correct API usage.

It's a good sign the tests get clearer too.