encapsulate not_yet_full() better
Unlike using the public API function addresses as sentinels, it is safe
to use a static, so this could be implemented as

    static void *not_yet_full(void) {
        return (void*)not_yet_full;
    }

but that means not_yet_full() would be instantiated and take up a few
instructions of code size, where with this not_yet_full() can always be
inlined and we still burn only that single byte.  This is the
archetypical "no big deal", but might as well.